Now makes use of AuthZ.

+++ main/modules/asset/delete.act.php
@@ -1,5 +1,9 @@
 <?
 
+// Check for our authorization function definitions
+if (!defined("AZ_DELETE"))
+	throwError(new Error("You must define an id for AZ_ACCESS", "concerto.collection", true));
+
 // Get the DR
 $drManager =& Services::getService("DR");
 $sharedManager =& Services::getService("Shared");
@@ -7,6 +11,22 @@ $assetId =& $sharedManager->getId($harmoni->pathInfoParts[3]);
 $asset =& $drManager->getAsset($assetId);
 $dr =& $asset->getDigitalRepository();
 $drId =& $dr->getId();
+
+// Check that the user can delete this asset
+$authZ =& Services::getService("AuthZ");
+$shared =& Services::getService("Shared");
+if (!$authZ->isUserAuthorized($shared->getId(AZ_DELETE), $assetId)) {
+	// Get the Layout compontents. See core/modules/moduleStructure.txt
+	// for more info. 
+	$harmoni->ActionHandler->execute("window", "screen");
+	$mainScreen =& $harmoni->getAttachedData('mainScreen');
+	$centerPane =& $harmoni->getAttachedData('centerPane');
+
+	$errorLayout =& new SingleContentLayout;
+	$errorLayout->addComponent(new Content(_("You are not authorized to delete this <em>Asset</em>."), MIDDLE, CENTER));
+	$centerPane->addComponent($errorLayout, MIDDLE, CENTER);
+	return $mainScreen;
+}
 
 // Delete the asset
 $dr->deleteAsset($assetId);
